requirements.txt
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.1
greenlet==3.0.3
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.9.1
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psycopg2-binary==2.9.9
pyparsing==3.1.2
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.15.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scikit-learn==1.5.1
scipy==1.14.0
six==1.16.0
SQLAlchemy==2.0.31
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.1
